Submission 2

+++ handlers/blog_delete_post_handler.py
@@ -7,6 +7,6 @@ class BlogDeletePostHandler(BlogHandler):
         key = db.Key.from_path('BlogPost', int(post_id),
                                parent=self.blog_key())
         p = db.get(key)
-        if p.user.key() == self.user.key():
+        if self.user and p and p.user.key() == self.user.key():
             db.delete(p)
         self.redirect('/blog/')
